Worked on acheiving novonix support
Worked on bringing novonix support to the mac version of the program. dVdQ and V-Q both appear to work but they have some quirks. The first cycle doesn't work with dVdQ, so for now I program into the program to skip the first cycle (not the long term solution). Additionally, the measured curves, while they do seem to match what I see in the previous version of the software, seem to be a lower resolution (they seem to be based on fewer data points). There is a lot to think about, but it seems very close to total integration.