Novonix works!
Novonix files are now supported (updated across all platforms) using a combination of Marc's parser and a workaround I (John) wrote to accommodate CSV file in the file uploader with streamlit. For some reason, CSV files weren't working with the file uploader so the memoryview was converted to a list of comma separated strings and the UniversalFormat was altered accordingly. Neware "cross-capacity" needs converting to a cumulative capacity. After this is done, there may be unit conversions that need fixing (the units seem wrong, but I can't know for sure until all curves are correct).